Fixed an issue that could cause the streams and stream ids to be incorrect due to the packet buffer having multiple stream change flags but only the capability to record one stream change Added in recording of the buffer level in ms

+++ src/demux_container.h
@@ -25,7 +25,7 @@ public:
   current_time(0),
   time_in_playlist(0) {};
   DemuxPacket *demux_packet;
-  int32_t current_time;
+  uint64_t current_time;
   double time_in_playlist;
   bool discontinuity;
 };
